feat(BA-1363): Remove `subscribed_actions` config, and change AuditLogReporter to AuditLogMonitor (#4400)
Backported-from: main
Backported-to: 25.6.5
Backported-of: 4400
