+++ changes/4400.misc.md
@@ -1,0 +1,1 @@
+Remove subscribed_actions config, and change AuditLogReporter to AuditLogMonitor.
